🚚 Restructure `callbacks` folder
