+++ banana/callbacks/__init__.py
@@ -1,6 +1,6 @@
-from .insert_row import InsertRowCallback
-from .load_main_table import LoadMainTableCallback
-from .load_side_menu import LoadSideMenuCallback
-from .open_history_modal import OpenHistoryModalCallback
-from .open_insert_modal import OpenInsertModalCallback
-from .update_cell import UpdateCellCallback
+from ._insert_row import InsertRowCallback
+from ._load_main_table import LoadMainTableCallback
+from ._load_side_menu import LoadSideMenuCallback
+from ._open_history_modal import OpenHistoryModalCallback
+from ._open_insert_modal import OpenInsertModalCallback
+from ._update_cell import UpdateCellCallback
