+++ banana/callbacks/_open_history_modal.py
@@ -5,9 +5,9 @@ from dash import html
 from dash_iconify import DashIconify
 import dash_mantine_components as dmc
 
-from ...core.tables import tables
-from ...core.history import LogType, get_history
-from ...core.utils import split_pathname
+from ..core.tables import tables
+from ..core.history import LogType, get_history
+from ..core.utils import split_pathname
 
 
 def t(text) -> str:
